Rename link linker to hyperlink linker and LTL accordingly
The kind of linker that is based on hyperlinks and can be plugged into
the linking pipeline before the actual linker is now called hyperlink
linker to allow differentiating between links as "linked entities" and
links as "hyperlinks". The link text linker is now called hyperlink
reference linker and the former link linker is now called hyperlinks
only linker.