Fix forkserver RAM usage issue
Previously, when using multiprocessing, RAM usage was roughly two times
higher than when using a single process. Mostly because with using the
forkserver multiprocessing method, the linking_system had to be imported
once in order to be usable in the link_entities.py but then also had to
be loaded with the multiprocessing.set_forkserver_preload() method.
To avoid loading this huge object twice, I now do a fake load when
importing it first (setting linking_system in forkserver_linking_system
to None) and only load it with the set_forkserver_preload() method
(by setting the configuration file accordingly). This reduces RAM usage
from roughly 90GB to roughly 55GB.